+++ NINA.Sequencer/Logic/Expression.cs
@@ -47,7 +47,7 @@ namespace NINA.Sequencer.Logic {
         public Action<Expression> Validator;
 
         public double? Default { get; set; }
-        public int[]? Range { get; set; }
+        public double[]? Range { get; set; }
         public bool IsExpression { get; set; } = false;
         public bool IsSyntaxError { get; set; } = false;
         public bool IsAnnotated {
